# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for ecommerce website",
        "title": "lapify eCommerce API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/admin/order/cancel/{orderid}": {
            "patch": {
                "description": "Cancels an order based on the provided order ID (for admin use).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Orders"
                ],
                "summary": "Cancel order (Admin)",
                "operationId": "admin-cancel-order",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Order ID to be canceled",
                        "name": "orderid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order cancelled successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/admin/order/details": {
            "get": {
                "description": "Retrieves the order details for admin based on pagination parameters.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Orders"
                ],
                "summary": "Retrieve order details for admin",
                "operationId": "get-admin-order-details",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number for pagination (default is 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items per page (default is 5)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order details retrieved successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/admin/order/update/{orderid}": {
            "patch": {
                "description": "Updates the status of an order based on the provided order ID and status (for admin use).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Orders"
                ],
                "summary": "Update order status (Admin)",
                "operationId": "admin-update-order",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Order ID to be updated",
                        "name": "orderid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "New status for the order",
                        "name": "status",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order updated successfully. Updated order status: {updated order status}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/admin/salesreport/date/{start}/{end}": {
            "get": {
                "description": "Generates a sales report based on the provided start and end dates.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Report"
                ],
                "summary": "Generate sales report by date range",
                "operationId": "sales-report-by-date",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start date for the report (format: 2-1-2006)",
                        "name": "start",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End date for the report (format: 2-1-2006)",
                        "name": "end",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sales report generated successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/admin/salesreport/payment/{start}/{end}/{paymentmethod}": {
            "get": {
                "description": "Generates a sales report based on the provided start and end dates and payment method.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Report"
                ],
                "summary": "Generate sales report by payment method and date range",
                "operationId": "sales-report-by-payment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start date for the report (format: 2-1-2006)",
                        "name": "start",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End date for the report (format: 2-1-2006)",
                        "name": "end",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Payment method for the report",
                        "name": "paymentmethod",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sales report generated successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/admin/salesreport/period/{period}": {
            "get": {
                "description": "Generates a sales report based on the provided period.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Report"
                ],
                "summary": "Generate sales report by period",
                "operationId": "sales-report-by-period",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Period for the report (e.g., 'monthly', 'quarterly', 'yearly')",
                        "name": "period",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sales report generated successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/order/payment/verify": {
            "post": {
                "description": "Verifies the payment for Razorpay based on the provided signature, Razorpay ID, and payment ID.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Orders"
                ],
                "summary": "Verify payment for Razorpay",
                "operationId": "verify-payment-razorpay",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Signature for payment verification",
                        "name": "sign",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Razorpay ID",
                        "name": "razorid",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Payment ID",
                        "name": "paymentid",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment successful. Invoice details: {invoice}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/user/order/cancel/{orderid}": {
            "patch": {
                "description": "Cancels an order based on the provided order ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Orders"
                ],
                "summary": "Cancel an order",
                "operationId": "cancel-order",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Order ID to be canceled",
                        "name": "orderid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order canceled successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/order/history": {
            "get": {
                "description": "Retrieves the order history for the authenticated user based on pagination parameters.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Orders"
                ],
                "summary": "Retrieve order history for the authenticated user",
                "operationId": "get-order-history",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number for pagination (default is 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items per page (default is 5)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order history retrieved successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "JWT": {
            "type": "apiKey",
            "name": "Authorise",
            "in": "cookie"
        }
    }
}